Add missing header and move misplaced enum
